# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/state-admin/dashboard": {
      "get": {
        "operationId": "StateAdminController_getStateDashboard",
        "parameters": [
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-state-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "State Admin overview & monitoring dashboard",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/departments": {
      "get": {
        "operationId": "StateAdminController_listDepartments",
        "parameters": [
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-state-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List departments in state",
        "tags": [
          "state-admin"
        ]
      },
      "post": {
        "operationId": "StateAdminController_createDepartment",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN or CENTRAL_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "State Admin creates department & appoints Department Head",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/departments/{id}": {
      "get": {
        "operationId": "StateAdminController_getDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get department details by ID",
        "tags": [
          "state-admin"
        ]
      },
      "put": {
        "operationId": "StateAdminController_updateDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update department details",
        "tags": [
          "state-admin"
        ]
      },
      "patch": {
        "operationId": "StateAdminController_patchDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update department details (PATCH alias)",
        "tags": [
          "state-admin"
        ]
      },
      "delete": {
        "operationId": "StateAdminController_deleteDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete department (fails if dependent records exist)",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/departments/{id}/status": {
      "patch": {
        "operationId": "StateAdminController_updateDepartmentStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Change department operational status (ACTIVE/SUSPENDED/INACTIVE)",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/departments/{id}/head": {
      "post": {
        "operationId": "StateAdminController_assignDepartmentHead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Assign or replace Department Head",
        "tags": [
          "state-admin"
        ]
      },
      "delete": {
        "operationId": "StateAdminController_removeDepartmentHead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Remove Department Head relationship",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/grievance-cells": {
      "post": {
        "operationId": "StateAdminController_configureGrievanceCell",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "STATE_ADMIN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigureGrievanceCellDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Configure department grievance cell & workflow",
        "tags": [
          "state-admin"
        ]
      },
      "get": {
        "operationId": "StateAdminController_listGrievanceCells",
        "parameters": [
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-state-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all grievance cells for a state (only for existing departments)",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/grievance-cells/{deptId}": {
      "get": {
        "operationId": "StateAdminController_getGrievanceCell",
        "parameters": [
          {
            "name": "deptId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get grievance cell configuration for department",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/analytics/revenue": {
      "get": {
        "operationId": "StateAdminController_getDepartmentRevenue",
        "parameters": [
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-state-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Department-wise revenue analysis",
        "tags": [
          "state-admin"
        ]
      }
    },
    "/api/v1/state-admin/analytics/kpis": {
      "get": {
        "operationId": "StateAdminController_getStateAnalytics",
        "parameters": [
          {
            "name": "stateId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-state-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "State Admin KPI overview",
        "tags": [
          "state-admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "UCSDP Backend API",
    "description": "Unified Citizen Service and Delivery Platform API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDepartmentDto": {
        "type": "object",
        "properties": {}
      },
      "ConfigureGrievanceCellDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}